comm

+++ src/ft_putadd/ft_putadd_f.c
@@ -6,13 +6,13 @@
 /*   By: hel-hadi <hel-hadi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/11 10:30:38 by hel-hadi          #+#    #+#             */
-/*   Updated: 2017/01/28 15:28:56 by hel-hadi         ###   ########.fr       */
+/*   Updated: 2017/01/29 14:20:08 by hel-hadi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../ft_printf.h"
 
-int		ft_nbr_add_len(int long nb)
+int		ft_nbr_add_len(size_t nb)
 {
 	int i;
