+++ src/ft_putuns/ft_putuns_maj_f.c
@@ -6,7 +6,7 @@
 /*   By: hel-hadi <hel-hadi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/10 13:21:49 by hel-hadi          #+#    #+#             */
-/*   Updated: 2017/01/25 10:27:55 by hel-hadi         ###   ########.fr       */
+/*   Updated: 2017/01/29 14:06:14 by hel-hadi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ int ft_putuns_maj_f(char *s, va_list *p)
 	diff = 0;
 	check_type = 1;
 	arg = va_arg(*p, unsigned long long);
-	nb = ft_count_num((int long)arg);
+	nb = ft_count_num(arg);
 	diff = ft_uns_size(s, check_type);
 	ft_handle_uns(s, check_type);
 	ft_size_chain_uns(s, nb, diff);
